+++ CHANGELOG.md
@@ -3,6 +3,8 @@ v0.7.0 (in development)
 - `properties2json` and `json2properties` now preserve the input ordering of
   keys by default; use the new `--sort-keys` option to get the old behavior
 - Drop support for Python 3.4
+- Gave `javaproperties`' `get` and `select` subcommands `--quiet` options for
+  suppressing warnings about missing keys
 
 v0.6.1 (2020-03-09)
 -------------------